% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstTimeInvariantMod.R
\name{EstTimeInvariantMod}
\alias{EstTimeInvariantMod}
\title{This function estimates time-invariant risk premia.}
\usage{
EstTimeInvariantMod(R, Rf, F_mat, dataset = "individual")
}
\arguments{
\item{R}{= matrix containing a possibly unbalanced panel of asset returns.}

\item{Rf}{= vector containing risk free returns.}

\item{F_mat}{= matrix containing factors (without constant).}

\item{dataset}{= character vector indicating if ExcR are "portfolio" ExcR or "individual" stocks ExcR.}
}
\value{
out = list of integer n, integer nchi(trimmed cross-sectional dimension),
vector a (intercept), matrix b (factor loadings), 3d-array CI (confidence intervals),
vector nu (annualized nu), vector RiskPremia (annualized risk premia).
}
\description{
This function estimates time-invariant risk premia.
}
