% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstTimeVariantMod.R
\name{EstTimeVariantMod}
\alias{EstTimeVariantMod}
\title{This function estimates time-varying risk premia.}
\usage{
EstTimeVariantMod(R, Rf, F_mat, Z, Zi, dataset = 3)
}
\arguments{
\item{R}{= matrix containing a possibly unbalanced panel of asset returns.}

\item{Rf}{= vector containing risk free returns.}

\item{F_mat}{= matrix containing factors (without constant).}

\item{Z}{= matrix containing the common instruments.}

\item{Zi}{= matrix containing the asset-specific instruments.}

\item{dataset}{= character vector indicating if ExcR are "portfolio" ExcR or "individual" stocks ExcR.}
}
\value{
out = list of integer n, integer nchi(trimmed cross-sectional dimension),
matrix beta (coefficients), 3d-array CI (confidence intervals),
matrix Table (annualized vec(F'), nu), vector RiskPremia (annualized),
vector nuT (annualized time-varying nu).
}
\description{
This function estimates time-varying risk premia.
}
