% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOS_data.R
\name{GOS_data}
\alias{GOS_data}
\title{GOS 2016 data preparation}
\usage{
GOS_data(dataset, K, model, p, q)
}
\arguments{
\item{dataset}{integer indicating the dataset to analyse.
1 = FF25, 2 = Indu44, 3 = CRSPCMST_ret. Each dataset contains:
R=returns, I=indicator for unbalanced characteristic, BM=book-to-mkt
Ti= time-series observations for each asset, T=TS size, n=CS size}

\item{K}{integer declaring the number of factors (mkt return, small minus big, high minus low, momentum).
default is 4.}

\item{p}{number of common instruments Z (1, 2 or 3 [default]).}

\item{q}{number of asset specific instruments Zi (0 or 1 [default]).}

\item{Model}{integer indicating the model. 1 (default) = time variant. 0 = time invariant.}
}
\description{
GOS 2016 data preparation
}
