% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOS_main.R
\name{GOS_main}
\alias{GOS_main}
\title{Reproduce GOS 2016}
\usage{
GOS_main(dataset = 1, K = 4, model = 0, p = 1, q = 0)
}
\arguments{
\item{dataset}{integer indicating the dataset to analyse.
1 = FF25, 2 = Indu44, 3 [default] = CRSPCMST_ret, 4 = tidy data. Each dataset contains:
R=returns, I=indicator for unbalanced characteristic, BM=book-to-mkt
Ti= time-series observations for each asset, T=TS size, n=CS size}

\item{K}{integer declaring the number of factors (mkt return, small minus big, high minus low, momentum).
default is 4.}

\item{p}{number of common instruments Z (1, 2 [default]).}

\item{q}{number of asset specific instruments Zi (0 [default] or 1).}

\item{Model}{integer indicating the model. 1 [default] = time variant. 0 = time invariant.}
}
\value{
out = list. If dataset = 1 or 2, out is a list of integer n, integer nchi(trimmed cross-sectional dimension),
vector a (intercept), matrix b (factor loadings), 3d-array CI (confidence intervals),
vector nu (annualized nu), vector RiskPremia (annualized risk premia).
If dataset = 3, out is a list of integer n, integer nchi(trimmed cross-sectional dimension),
matrix beta (coefficients), 3d-array CI (confidence intervals),
matrix Table (annualized vec(F'), nu), vector RiskPremia (annualized),
vector nuT (annualized time-varying nu).
}
\description{
Reproduce GOS 2016
}
