% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOS_tidy_input.R
\name{GOS_tidy_input}
\alias{GOS_tidy_input}
\title{This function is project specific and it transfroms specific tidy data into wide data for reproducing GOS 2016.}
\usage{
GOS_tidy_input(asset_dataf = lhs_data, Rf = NULL, factor_dataf = rhs_data,
  model = 0, instument_dataf = tv_coeff_data, dataset = "individual")
}
\arguments{
\item{asset_dataf}{tidy dataframe containing assets information.
it must have (at least) the following columns: date, value (total returns),
hash_key_entity (uniquely identifying the assets).}

\item{Rf}{NULL for this project. Ideally it would be a dataframe containing risk free information.}

\item{factor_dataf}{tidy dataframe containig factors information.
it must have (at least) the following columns: date, value, hash_key_entity (uniquely identifying the factors).}

\item{instrument_dataf}{tidy dataframe containing information on the instrumental variables
for time varying coefficients. If model = 0, is set to NULL. If model = 1,
it must have the following columns: date, value, hash_key_entity (uniquely identifying the instruments).}
}
\value{
wide data ready for estimation.
}
\description{
This function is project specific and it transfroms specific tidy data into wide data for reproducing GOS 2016.
}
